requirements.txt
appdirs==1.4.3
argcomplete==1.10.0
colorama==0.4.1
Pillow==6.1.0
pyserial==3.4
python-escpos==2.2.0
pyusb==1.0.2
PyYAML==5.1.2
qrcode==6.1
six==1.12.0
